requirements.txt
tabula-py==2.9.0
pandas>=2.0.0
pdfplumber>=0.11.0
requests>=2.32.0
